[JBWS-3733]:Fix ws subystem schema;Merge WebServiceEndpointMetricsTestCase with WSTestCase
